refactor(app): extract prompt store types and editor DOM serialization
Move PromptStore/HistoryStore types to prompt-input/store-types.ts and
DOM <-> Parts bidirectional serialization (createPill, isNormalizedEditor,
renderPartsToEditor, parseEditorToParts) to prompt-input/editor-serialize.ts.

Pure functions, no Solid reactivity. Pairs with editor-dom (cursor primitives).
First step of splitting prompt-input.tsx (1602 lines) into business-domain modules.
